Improve UI

+++ src/gui_interface.h
@@ -433,8 +433,8 @@ public:
 
     bool config_file_exists = true;
 
-    int link_quality_ = 0; // Percentage
-    int packet_loss_ = 0;  // Percentage
+    float link_quality_ = 0; // Percentage
+    float packet_loss_ = 0;  // Percentage
     int drone_fec_level_ = 0;
 
     // Use gstreamer for decoding instead of ffmpeg
